Reworked and optimized the AME Catalogue script
Did some optimization of the script, and condensed many sections into repeatable functions. Also, added text-to-speech functionality, so the program can be run in the background and the progress and results will be read to you by the charming voice of Microsoft Sam. This last feature is untested in anything other than Windows 10, however, since it's a built-in function of the OS, it should work in any Windows version.