Added additional files needed for added functionality in the AME Catalogue script.
The setup file should be run every time custom object names are added, or before you run the main R script. The subset script is for picking apart and subsetting the data produced by the main R script. The batch file and VBS script are necessary for the text-to-speech functionality. Check the Readme for more info on this functionality.